pyproject.toml
[project]
name = "wardrobe-ai-hybrid"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.135.1",
    "google-genai>=1.68.0",
    "langchain>=1.2.13",
    "langchain-core>=1.2.20",
    "langchain-google-genai>=4.2.1",
    "langchain-google-vertexai>=3.2.2",
    "langgraph>=1.1.3",
    "llama-index-core>=0.10.0",
    "llama-index-embeddings-fastembed>=0.3.5",
    "llama-index-vector-stores-qdrant>=0.6.1",
    "python-dotenv>=1.0.0",
    "qdrant-client<1.13",
    "transformers[torch]>=5.3.0",
    "uvicorn>=0.42.0",
]

[[tool.uv.index]]
url = "https://private.hubteam.com/pypi/simple/"
default = true


[[tool.uv.index]]
url = "https://pypi.org/simple/"
